validations

+++ backend/BookStore/Controllers/BooksController.cs
@@ -4,6 +4,7 @@ using BookStore.Core.Models;
 using FluentValidation;
 using FluentValidation.Results;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
@@ -11,6 +12,7 @@ namespace BookStore.Controllers
 {       
     [ApiController]
     [Route("[controller]")]
+    [EnableCors("AllowSpecificOrigin")]
     public class BooksController : ControllerBase
     {
         private readonly IBooksService _booksService;
@@ -23,7 +25,6 @@ namespace BookStore.Controllers
 
         [HttpGet]
         [Authorize(Policy = "User")]
-        /*RequireAuthorization(policy => policy.AddRequirement(new PremissionRequirement[Permission.Read]))*/
         public async Task<ActionResult<List<BooksResponse>>> GetBooks()
         {
             var books = await _booksService.GetAllBooks();
@@ -37,10 +38,11 @@ namespace BookStore.Controllers
         {
             ValidationResult results = _validator.Validate(request);
 
-            if (! results.IsValid)
+            if (!results.IsValid)
             {
                 return BadRequest(results.Errors);
             }
+
             var (book, error) = Book.Create(
                 Guid.NewGuid(),
                 request.Title,
@@ -59,6 +61,18 @@ namespace BookStore.Controllers
         [Authorize(Policy = "Admin")]
         public async Task<ActionResult<Guid>> UpdateBook(Guid id,[FromBody] BooksRequest request)
         {
+            if (!await _booksService.IsBookExist(id))
+            {
+                return BadRequest("Book does not exist");
+            }
+
+            ValidationResult results = _validator.Validate(request);
+
+            if (!results.IsValid)
+            {
+                return BadRequest(results.Errors);
+            }
+
             var bookId = await _booksService.UpdateBook(id, request.Title, request.Description, request.Price);
             return Ok(bookId);
         }
@@ -67,6 +81,11 @@ namespace BookStore.Controllers
         [Authorize(Policy = "Admin")]
         public async Task<ActionResult<Guid>> DeleteBook(Guid id)
         {
+            if (!await _booksService.IsBookExist(id))
+            {
+                return BadRequest("Book does not exist");
+            }
+
             var bookId = await _booksService.DeleteBook(id);
             return Ok(bookId);
         }
